scripts/hadith_import/requirements.txt
requests>=2.25.1
pymysql>=1.0.2
python-dotenv>=0.19.0
